Changes to support Ansible Collection SDK

+++ sdkAutomator/executePythonResources.py
@@ -14,7 +14,8 @@ class executePythonResources():
     config_file = os.getcwd() + '/oneview-python/examples/config.json'
 
     def __init__(self, resource_dict):
-        shutil.copyfile(self.config_rename_file, self.config_rename_dummy_file)
+        if (os.path.isfile(self.config_rename_file)):
+            shutil.copyfile(self.config_rename_file, self.config_rename_dummy_file)
         self.resource_dict = resource_dict
         self.exe = self.load_resources()
         self.generate_config_values()
